+++ itesa/middleware.js
@@ -18,5 +18,5 @@ export async function middleware(request) {
 }
 
 export const config = {
-  matcher: [],
+  matcher: ["/homeUser"],
 };
