+++ itesa/pages/homeuser.js
@@ -1,9 +1,13 @@
 import React from "react";
 import WalletCard from "../components/WalletCard";
+import Persistence from "../components/Persistence";
+import { useSelector } from "react-redux";
 
 const HomeUser = () => {
+  
   return (
     <div>
+      <Persistence />
       <WalletCard />
     </div>
   );
